Hack to ensure valid dynamic EnumProperty values
There doesn't seem to be a proper way to ensure EnumProperty values are always a value on the enum_items list. For instance if an EnumProperty is dynamicaly generated from a collection property, if one of those collection property items are removed, the EnumProperty will not change, resulting in an invalid enum value. This commit just validates all the enum values whenever items are removed from the collection properties.