Auto-calculate bone tag setting per bone
Removes "Auto Calculate Bone Tag" export setting.

Now each bone has a setting to choose between automatically
calculated tag based on the bone name and a manually set tag. When
setting a tag (either during import or by the user), if the value
matches the auto-calculated value, it is kept as auto-calculated.
Otherwise, it is set as manual tag.

`BoneProperties.tag` directly returns the correct tag value, so the
caller doesn't need to worry about auto-calculation settings, reducing
the chance of bugs due to forgetting to check the auto-calculation.

Fixes #751.