feat(ytyp): add gizmos for archetype extensions
Adds new workspace tool "Edit Archetype Extensions" that when active
with a YTYP selected, it shows gizmos for all the extensions of its
archetypes with a linked asset. These gizmos can be selected and
interacted with to edit the archetype extensions.

Also added "duplicate extension" operators for both archetype and
entity extensions.

Some shortcuts are available when the tool is active:
- DEL: delete the selected archetype extension.
- SHIFT+D: duplicate the selected archetype extension.

Closes #472.
Closes #685.