fix(shader): use correct UV map per texture sampler
UV map indices are defined in Shaders.xml with the `uv` attribute in
the texture parameters.

When creating the shader material, it will create UV map nodes for all
UV maps used by the shader and connect them to the corresponding texture
nodes, if they haven't been connected already.

If the `uv` attribute is missing, it won't be connected to any
UV map node. Used for textures that are not sampled with UV maps.

Fixes #711.