fix(ydr): correctly set Capsule type to created lights
Previously, it was setting `Light.light_properties.type`, which doesn't
exist, instead of `Light.sollum_type`.

Spot and Point lights worked correctly because the getter of
`Light.sollum_type` checks the blender light type but Capsule needs an
additional flag as it re-uses point lights, which was not set as the
correct property wasn't used.

Fixes #784.