fix(ydr): assign correct default render bucket when creating shader materials
The shaders XML didn't specify which render bucket was used for each
shader preset (.sps), only a list of render buckets. So when creating
a material it chose the first one, which most of the time is incorrect.

Now the XML has the render bucket for each shader preset and assigns
them correctly when creating the material.