feat(ydr): light culling plane gizmo
Lights with culling plane enabled show a gizmo representing the culling
plane. This gizmo is a round plane with a cross pattern and arrows to
indicate which side of the culling plane has light. It can be interacted
with to move the culling plane along its normal (change its offset).

To change the culling plane normal, there is a new workspace tool that
will show a rotation widget to modifies the normal on the selected
light. The tool UI also has buttons to flip the plane or set its normal
to the XYZ axes.

Refactored a bit of the existing capsule light gizmos code to better
support culling plane gizmos along with capsules. Capsule and culling
plane gizmos can also now be clicked on to select the light object.

Closes #786.