fix(yft): incorrect group parent index when direct parent bone has no physics enabled
When a parent bone didn't have physics enabled but the child bone did,
the frag group for the child bone would be exported as a root group
with no parent (parent index = 255). Instead, it should search for a
grandparent bone with physics enabled first and consider it its parent
in the frag group.

This could cause crashes, specially on vehicles, which expect a single
root group, the chassis.

Fixes #841.