chore(git): exclude unnecessary files from git archives
These files won't be included in `git archive`, making the output
archive size smaller, which increased quite a bit due to the assets in
`tests/`.

These files are not needed to use the addon in Blender, only for
development or documentation. With this we avoid more manual steps
when preparing a new release. This also affects the "Download ZIP"
button from GitHub.