fix(ydr): LOD meshes attached to wrong bones on import
The import code assumed that the DrawableModels in the arrays for the
different LODs were in the same order based on the bone index.

For example:
- High LOD array has models with bone indices [0, 1, 2].
- Med LOD array has a single model with bone index [0].
  - Then, import would create a drawable model with 2 LODs attach to
    bone 0.
- If instead, the High LOD array was ordered like [2, 1, 0].
  - Now, import would attach the model in the Med LOD array to bone 2,
    even though it is for bone 0.

The game doesn't require this (and we aren't ensuring that on export
anyways). Instead, we now connect the models in the different LODs
taking into account the bone index, instead of just the order.

Fixes #876.