fix(ytyp): support entity set entities attached to portals
In the entity set locations array, the MSB of a value indicates whether
it refers to a portal index or a room index. We only considered them as
room indices. Now, we check the MSB on import and correctly set it on
export.

An example .ytyp affected by this was `vw_vwdlc_int_02.ytyp`.