fix(ydr): merge vertices with same position on import
We were creating a Blender vertex per game mesh vertex, which resulted
in faces using the same vertex position but with different attributes
not being connected.

We don't need to create a vertex per game vertex. Vertices with the same
position can be combined into one vertex with attributes set in the mesh
loops using that vertex.