feat(ydr): support cable meshes
On import, we convert the game mesh to a simplified mesh with only
vertices and edges, without faces. On export, we rebuild the topology
like the game expects it.

The mesh has user-editable vertex attributes: radius, diffuse factor,
micromovements scale, phase offset and material index.

Material index is only required when the same mesh uses multiple
materials (only 25 vanilla models do this). We do it through a custom
attribute because Blender materials are assigned to faces, which we
don't have in cable meshes.