fix(ydr): export correct light position when `matrix_parent_inverse` is set
When `matrix_parent_inverse` is set, `location` is not what we want
(i.e. coordinates relative to parent). Also, using `matrix_world`
instead of `matrix_local` because lights object may not be a direct
child of the parent object (e.g. on import we create a wrapper object
for light, which could end up transformed in some way by the user).

Fixes #865.