fix(ui): shaders list empty on new scene
Moved the collection property to the WindowManager, which is a singleton
so the same instance is always available.

Also allows us to move the shader collection initialization to
`register()` instead of waiting for a .blend file to load, on a
`load_post` handler.

Fixes #609.

Not done here but other collections that exist only for UI purposes
should be moved to WindowManager as well, the preset collections come to
mind.