fix(lods): imported models not defaulting to very-high LOD when available
`LODLevels.set_highest_lod_active` was changed to iterate over the
`LODLevel` enum instead of a explicit list in a previous commit. But in
the enum, VERYHIGH was last. Reordered the enum so it is sorted from
highest to lowest.

Replaced some other hardcoded explicit lists with the enum as well.