fix(yft): export errors when bones have .00# suffix
When creating physics groups, it was removing the suffix from the name,
later causing errors by confusing the bones with the "same" name.

Bone names are local to the armature, so doesn't really make sense to
remove the suffix. And the code that creates the skeleton doesn't remove
them either anyways.