fix: use COPY_TRANSFORM constraint instead of CHILD_OF for object-bone parenting
Blender 4.2 broke how we were using CHILD_OF constraint. Seemingly, we
were using CHILD_OF wrong. CHILD_OF adds the offset after applying the
parent transform, so it is kinda moving the object twice. It is meant
for objects not in the same hierarchy.

Before Sollumz kinda hacked around it by setting the target/owner spaces
which all constraints have, but that's not even shown in the UI by
default (Sollumz added it) so I doubt it is even properly supported by
Blender.

Instead, we now use the COPY_TRANSFORM constraint which just uses the
bone transform directly. Most properties are the same as CHILD_OF so it
doesn't need many changes in the code.

Added versioning to convert CHILD_OF constraints to COPY_TRANSFORM.