chore: add extension manifest
Fixed some issues that caused the add-on to fail to install as a
extension.

Updated the prepare_release.ps1 to take into account the
blender_manifest.toml.

The new manifest allows us to use any version string following SemVer,
instead of just numbers. Therefore, for development builds (any commit
without a 'v*.*.*' tag), from now on the version will include the commit
hash to help us differentiate them from releases. This is done via a
placeholder marker which `git archive` will fill out (note: we cannot
use `blender --command extension build` for this because it won't
replace the placeholder, but the ZIP generated by `git archive` works
fine as an extension).

Additionally, development builds have the extension ID `sollumz_dev`
while releases are `sollumz`. This allows us to have both packages in a
extensions repository (for latest commit and for latest release).

The prepare_release.ps1 script now creates two commits. First one with
just the release version and ID. This is the commit that is tagged. A
second one adding back the commit placeholder and development ID.

Closes #931.