tweak(obb): rework oriented bounding box calculations
The axis parameter was filled with the same value for x and y, causing
the bounding box to always by oriented wrong if the object was turned
in either the x or y axis.

This implementation still uses brute force, but estimates the angle
quite well and within 0.2 seconds. Uses Fibonnaci lattice for axis
selection.

Also added settings to select number of samples used and number of
angles it tries. Added REGISTER option to operator to tweak these
settings in the redo panel.

Set a minimum on axis volume. When selecting faces on a plane the
volume would become zero making it pick whichever box it found first,
which more often then not was not the optimum shape.