fix(yft): paint layer now backed by 'matDiffuseColor' parameter
Before it was a separate property, and 'matDiffuseColor' was overwritten
during export. Now, the property uses getters and setters to read/write
the value of 'matDiffuseColor', so this parameter is the source of truth
now.

Additionally, added the missing Pearlescent (3) and Default (5) paint
layers, renamed Not Paintable to Custom, and updated the descriptions.

This also fixes a bug where the 'matDiffuseColor' was only overwritten
in the non-hi .yft.xml, the hi one still had the value displayed in the
UI.

Fixes #997.