fix(shader): incorrect display of some tint palette pixels
Due to the sRGB -> linear -> sRGB conversion we have to do on the color
attribute values we lose precision and the UVs of pixels 10-12 don't
quite get to the pixels they should be in:
```
Index  UV               Sampled Pixel
  9    0.035294 * 256 =  9.035
 10    0.039043 * 256 =  9.995
 11    0.042951 * 256 = 10.9955
 12    0.046868 * 256 = 11.9982
 13    0.050786 * 256 = 13.0012
```

So after obtaining the UV, we round it as follows `uv.x = round(uv.x * width) / width`

Fixes #1029.

Cleaned up a bit the geometry nodes too: removed use of deprecated
`ShaderNodeCombineRGB` and UV capture attribute data type is now vector
instead of color.