fix(shader): broken cutout hard edges in Blender 4.2+
Since Blender 4.2, the "Alpha Clip" blend method we were using for
cutout shaders is no longer supported. Now, we have to do the alpha
test ourselves in the shader node tree.

As a bonus, the "HardAlphaBlend" parameter is now taken into account in
the cutout shaders that have it.

Fixes #1002.