feat(shader): support textures in shader presets
Texture names are now saved in shader presets. When applying a preset,
the material textures are overwritten by those in the preset. The image
is first searched in the .blend loaded images. If not found, the shared
textures directories are searched.

There is a new preference to enable or disable this behaviour when
applying a shader preset.

Closes #1027.