fix(ytyp): only select linked object when archetype/entity was changed through the UI list
We don't always need to select the linked object. For example, in the
archetype extension gizmos, clicking on a gizmo may change the current
archetype index but we don't want to make the object active.

Add new properties that wrap the archetype/entity indices to explicitly
opt-in into selecting the linked object.