perf(ybn): revert margin calculation
`shrink_mesh` is too slow in some cases. The exact value of margin is
not important enough to justify the increased export times. So go back
to a hardcoded margin for geometry bounds, and actually use the correct
minimum this time, 0.025 instead of the 0.0025 we were using before.

This calculation was introduced as part of a fix (77b718725e15f5d77717fd8904b4ea86c4723782)
but was not relevant to the actual issue. Just larger bounding boxes
were needed. It is still fixed when using a hardcoded margin.