fix(ybn): centroid calculation hangs during export
Miniball expects no duplicated points but we were not ensuring it. This
can cause repeated failures, and since we had an infinite loop trying
again and again, it freezed the export process.

Now, pass only the unique vertices and if any error still occurs,
fallback to the sphere enclosing the bounding box.