fix(ydr): correct handling of non-color data colorspace
When using a custom OpenColorIO config or from other software, the
colorspace for non-color data may not be called "Non-Color". This is
only the name in the default Blender OCIO config. Instead, now we use
whatever colorspace is assigned to the data role in the config (through
the `is_data` property).