feat(ynv): hacks to allow import/export without losing data
Should be good enough for editing flags. Editing the mesh itself will
mess it up, particularly adding or removing polygons. Edge adjacency is
not calculated at all anymore, instead the indices from import are
stored in an attribute.

Import keeps the mesh as close as possible to the original file data,
it does not try to fix the mesh anymore, not even merging vertices, so
every polygon is loose now. Only fix is adding an extra vertex to the
DLC stitch polygons which contain only 2 vertices, so Blender can create
the face and not ignore it. This vertex is removed on export.

Requires unreleased CodeWalker version (not dev48, newer).