feat(vertex_paint): isolate channels and paint gradient tools
Introduces two new vertex color tools:
- Isolate color channel: allows to visualize and edit RGBA channels
  independently.
- Paint gradient between two colors, based on Vertex Color Master
  gradient tool.

Both accessible only in Vertex Paint mode. Isolate channels is in a new
N-panel called "Vertex Paint" and in a pie menu, binded to Shift+T by
default. Gradient is accessible as a workspace tool or in the Paint menu.

Moved the old Terrain Painter panel to the new N-panel and pie menu.