fix(ydr): incorrect culling plane gizmo on rotated entities
The culling plane gizmo rotation from the normal always used global
coordinates and didn't take into account the parent object rotation so
it was placed incorrectly in some cases, for example, in MLO entities
with non-identity rotation.

Now, the plane normal is relative to the parent object rotation and
gizmo is correctly rotated.