feat: binary asset import/export, shattermap generation, szio library
This a big one. A squashed commit from several months of work.

Main change is support for importing and export binary assets, both from
Legacy (Gen8) and Enhanced (Gen9) game version. CWXML is still
supported. This involved a large refactoring of the import and export
code, mainly in *_io.py files now. The previous CWXML-specific system is
still available under a "Legacy Import/Export" toggle in the user
preferences in case users encounter a showstopper bug in the new system.
This legacy system will be removed in a future version.

Code that interacts with game formats has been moved to a
Blender-independent library called `szio`. This includes the `cwxml/`
module which is no longer in this repo, but available through `szio`.

`szio` provides a format-agnostic interface so we can treat binary
formats and CWXML formats in the same way in the code that interacts
with Blender. Binary formats supports is provided by the PyMateria
library developed by Cfx.re.

Due to these dependencies and that we cannot redistribute PyMateria
there is a new "Install Dependencies" dialog upon the first
installation of Sollumz. Here you can choose whether you want to install
PyMateria or not, as it is an optional dependency. `szio` is always
required.

With PyMateria installed, we also support automatic vehicle window
shattermap generation on export. There a new button "Preview Windows" to
see the generated shattermaps. The manual shattermaps are still
supported but not recommended. On import, shattermaps are no longer
added to the scene unless the "Import Window Shattermaps" import setting
is enabled.

Also added name tables support to resolve hashed names. User can add
name table files from the preferences and will be used to resolve hashes
when importing assets.

Finally, there are other small changes, like new the 'Fragment > Vehicle
Tools' subpanel to combine the tools related to vehicles, and probably a
few more.