perf(ydrexport): optimize single LOD + many models drawable export
Changing the LOD level during export changes the object mesh which
causes the depsgraph to be rebuilt. This is generally fast but on large
scenes it can take longer (e.g. complex interiors, ~11K Blender objects,
~0.5s to rebuild the depsgraph).

The depsgraph is used in  `create_model` > `get_evaluated_obj`. When
exporting drawables, the depsgraph rebuild happens `numDrawableModels * numLODsPerModel`
times, even if there is a single LOD on each drawable. Normally, not
many drawable models are used so this is not a huge issue but with many
models it slowdowns the export significantly. On a drawable with 187
models and a single LOD, from ~105s to ~10s export time with this
commit, and builds the depsgraph once instead of 187 times.

No improvement in the multiple LODs case and there are probably other
cases where the depsgraph is rebuilt unnecessarily/too many times, like
exporting multiple separate drawables at once.